Specify DJANGO_SETTINGS_MODULE in supervisord conf
Updates the supervisord config to explicitly set
`DJANGO_SETTINGS_MODULE` to the graphite settings module.

This appears to be necessary (in some cases) since changing from
the apt-provided Django to the pip-provided Django, so that Django
can find the correct settings. If this is not set then graphite
web will refuse to load leaving `ImproperlyConfigured` exceptions
in `/var/log/supervisor/graphite-webapp.log`.