Scale the carbon backend (#11)
* Update to grafana 6.4.5

* Don't require login to add dashboards

By default, we have to login with the admin/admin password, and
immediately re-type admin/admin to confirm the new password.  Really
what we want is a login-less ability to CRUD dashboards.

* Increase number of http workers

The default carbon-cache configuration uses a single Python thread,
which can quickly become CPU-bound, and not disk-IO bound.  By scaling
to 8 processes, we make it possible to scale better if the observer node
has more than 1 vCPU and substantial disk bandwidth.

Tests showed that with more than 2-4 carbon-cache processes, the Python
carbon-relay became the bottleneck.  Hence, Received packets are relayed
to carbon-cache processes by a C implementation of carbon-relay.